requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-apache-spark
apache-airflow-providers-google
google-cloud-storage
pyspark>=3.1.3
pandas>=1.0.5
pyarrow>=4.0.0
psycopg2-binary
apache-airflow-providers-postgres>=3.0.0
